updated project

+++ stock-table.php
@@ -25,7 +25,7 @@
       <th>Price</th>
       <th>Quantity</th>
       <th>Product ID</th>
-      <th>Options</th>
+      <th colspan="2">Options</th>
     </tr>
   </thead>
 
@@ -50,7 +50,8 @@
           <td>".$result['Price']."</td>
           <td>".$result['Quantity']."</td>
           <td>".$result['Product_ID']."</td>
-          <td><a href='delete.php?id=$result[Product_ID]'>Delete</td>
+          <td><a href='modify.php?brand=$result[Brand]&product=$result[Product_Name]&color=$result[Color]&size=$result[Size]&price=$result[Price]&quantity=$result[Quantity]&id=$result[Product_ID]'>Modify</td>
+          <td><a href='delete.php?id=$result[Product_ID]' onclick='return confirmDelete()'>Delete</td>
           </tr>
           ";
         }
@@ -59,6 +60,7 @@
       echo "No records found";
     }
   ?>
+  
   </tbody>
 
   <tfoot>
@@ -70,12 +72,65 @@
     <th></th>
     <th>Total</th>
     <th id="result"></th>
-    <th></th>
+    <th colspan="2"></th>
     </tr>
   </tfoot>
   </table>
 
+<script>
+  function confirmDelete() {
+    return confirm('Do you really want to Delete this row?');
+  }
 
+  /**
+ * Sorts a HTML table.
+ *
+ * @param {HTMLTableElement} table The table to sort
+ * @param {number} column The index of the column to sort
+ * @param {boolean} asc Determines if the sorting will be in ascending
+ */
+function sortTableByColumn(table, column, asc = true) {
+  const dirModifier = asc ? 1 : -1;
+  const tBody = table.tBodies[0];
+  const rows = Array.from(tBody.querySelectorAll("tr"));
+  
+  // Sort each row
+  const sortedRows = rows.sort((a, b) => {
+    let aColText = a.querySelector(`td:nth-child(${column + 1})`).textContent.trim();
+    let bColText = b.querySelector(`td:nth-child(${column + 1})`).textContent.trim();
+  
+    if (!isNaN(parseFloat(aColText)) && !isNaN(parseFloat(bColText))) {
+      aColText = parseFloat(aColText);
+      bColText = parseFloat(bColText);
+    }
+  
+    return aColText > bColText ? (1 * dirModifier) : (-1 * dirModifier);
+  });
+  
+  // Remove all existing TRs from the table
+  while (tBody.firstChild) {
+    tBody.removeChild(tBody.firstChild);
+  }
+  
+  // Re-add the newly sorted rows
+  tBody.append(...sortedRows);
+  
+  // Remember how the column is currently sorted
+  table.querySelectorAll("th").forEach(th => th.classList.remove("th-sort-asc", "th-sort-desc"));
+  table.querySelector(`th:nth-child(${ column + 1})`).classList.toggle("th-sort-asc", asc);
+  table.querySelector(`th:nth-child(${ column + 1})`).classList.toggle("th-sort-desc", !asc);
+}
 
+document.querySelectorAll(".table-sortable th").forEach(headerCell => {
+  headerCell.addEventListener("click", () => {
+    const tableElement = headerCell.parentElement.parentElement.parentElement;
+    const headerIndex = Array.prototype.indexOf.call(headerCell.parentElement.children, headerCell);
+    const currentIsAscending = headerCell.classList.contains("th-sort-asc");
+    
+    sortTableByColumn(tableElement, headerIndex, !currentIsAscending);
+  });
+});
+
+</script>
 </body>
 </html>
